+++ src/features/mm_bone_data_feature.h
@@ -44,7 +44,7 @@ public:
 
     virtual int64_t get_dimension_count() const override;
 
-    virtual PackedFloat32Array bake_animation_pose(Ref<Animation> p_animation, float time) const override;
+    virtual PackedFloat32Array bake_animation_pose(Ref<Animation> p_animation, double time) const override;
 
     virtual PackedFloat32Array evaluate_runtime_data(const MMQueryInput& p_query_input) const override;
